CASMINST-3669: check sp2 repo to determine Nexus health
At some point in the CSM 1.2 development timeline an SP3 repo was
configured in Nexus.  This appears to no longer be the case.  And
since the RPMs we're installing still live in a repo with sp2 in
the name, check for the existence of the sp2 repo in order to
determine Nexus health.

While we're here, adjust the warning message to be more helpful
in the event Nexus is deemed unhealthy.