MTL-1682 Move kernel provisioner
The kernel provisioner lived in the wrong directory, which was a minor
aesthetic issue. The main change here ensures that the kernel is not
locked before csm-rpms is invoked. The issue was that if a newer kernel
was defined in csm-rpms that it would fail to install due to the zypper
lock.

MTL-1682 Remove redundant kernel version

We lock the kernel with csm-rpms, this extra kernel definition is
entirely unnecessary and confusing to maintaining the kernel.

This also drastically cleans up the provisioners for the ncn-common
layer so they more clearly define contexts.