CASMINST-5122 Remove `/usr/local/bin/pip3`
Python3.6 installs `/usr/bin/pip3`, and Python 3.9 is installing
`/usr/local/bin/pip3`. The way the `$PATH` is currently set is causing
`/usr/local/bin/pip3` to take precedence, but this is confusing because
`python3` points to `python3.6` and `pip3` does not point to a Python3.6
install in this case.

By removing `/usr/local/bin/pip3`, the default `python3` and `pip3`
commands will align to the same version of Python.