MTL-1858 Tidy-up and Acknowledge All Tests
Changes so all tests are acknowledged, previously failures in non-common
tests were not flunking the build.

This also removes some cruft, such as:
- The extra call to `create-kis-artifacts.sh` in the ncn-common layer,
  since its artifacts were not uploaded nor used.
- The lingering `config.xml` file in pit-common which is not used at
  all.

The new metal-provision will consolidate ncn-common-setup and
ncn-common-install together, and in doing so this change invokes
metal-provision *after* all RPMs are installed. This enables us to
choose different version of Python for our Ansible, which we couldn't do
previously because we ran Ansible prior to installing other Python
versions.

This also formats all the packer files, and adds Licenses to any code
this PR touched.