MTL-1910 Disable NCN Common Rebuild
This change disables the need to rebuild the common layer every build,
by moving the `pre-install-toolkit` out of the common layer into our NCN
layer.

By moving `pre-install-toolkit` out of the common layer, we no longer
depend on building the common layer as a step in the build for
constructing all of our target artifacts. By moving
`pre-install-toolkit` into the final layer, all we need to build now is
the last layer. This shaves off ~30minutes of build time.

This includes some ancillary fixes:
- Fix local-builds of the Base Layer, use the correct SP3 ISO when
  Jenkins isn't overriding the `source_iso_uri` argument (e.g. local
  builds)
- Also allow re-runs on git-tags; use `try/catch` to allow failures
- Cleanup licenses and `#!/usr/bin/env bash` shebangs
- CASMINST-5256 use writeable env vars for release tags since the
  `release()` call does not use the Jenkins-plugin and depends on env
  vars.