Fix image pulls and versions
The `pre-instal-toolkit` needed to key off of
`params.rebuildCommonLayer` but was using `params.rebuildBaseLayer`,
this was causing it to not setup the right parameters for Google and
Metal.

Additionally this change fixes the image pulls, the logic here has been
broken for some time but broken further by MTL-1910. This helps
debugging by showing the correct original artifact Now it will:
- When skipping layers, the downloaded artifact will contain the correct
  version identifier.
- When not skipping layers, the previous layers artifacts are used.
- When skipping layers and not using latest, the pulled artifact will
  use the desired base version.

Before the above bullets were fixed in this PR, the downloaded artifact
was named with the current build's version - this was misleading.